Extract common code from broker jobs
[finishes #169365149]
[PR]
