+++ app/jobs/v3/services/synchronize_broker_catalog_job.rb
@@ -1,3 +1,5 @@
+require 'jobs/v3/services/service_broker_catalog_updater'
+
 module VCAP::CloudController
   module V3
     class SynchronizeBrokerCatalogJob < VCAP::CloudController::Jobs::CCJob
@@ -37,30 +39,18 @@ module VCAP::CloudController
 
       class Perform
         def initialize(broker_guid)
-          @broker_guid = broker_guid
           @broker = ServiceBroker.find(guid: broker_guid)
-          @formatter = VCAP::Services::ServiceBrokers::ValidationErrorsFormatter.new
-          @service_event_repository = VCAP::CloudController::Repositories::ServiceEventRepository::WithBrokerActor.new
-          @client_manager = VCAP::Services::SSO::DashboardClientManager.new(broker, service_event_repository)
-          @service_manager = VCAP::Services::ServiceBrokers::ServiceManager.new(service_event_repository)
+          @catalog_updater = VCAP::CloudController::V3::ServiceBrokerCatalogUpdater.new(@broker)
         end
 
         def perform
           synchronizing_state
 
-          catalog = VCAP::Services::ServiceBrokers::V2::Catalog.new(broker, broker_client.catalog)
+          warnings = @catalog_updater.refresh
 
-          raise fail_with_invalid_catalog(catalog.validation_errors) unless catalog.valid?
-          raise fail_with_incompatible_catalog(catalog.incompatibility_errors) unless catalog.compatible?
+          set_to_available_state
 
-          unless client_manager.synchronize_clients_with_catalog(catalog)
-            raise fail_with_invalid_catalog(client_manager.errors)
-          end
-
-          service_manager.sync_services_and_plans(catalog)
-
-          available_state
-          collect_warnings
+          warnings
         rescue
           failed_state
           raise
@@ -68,12 +58,10 @@ module VCAP::CloudController
 
         private
 
-        attr_reader :broker_guid, :broker,
-          :formatter, :client_manager, :service_event_repository,
-          :service_manager, :warnings
+        attr_reader :broker, :warnings
 
-        def broker_client
-          @broker_client ||= VCAP::Services::ServiceClientProvider.provide(broker: broker)
+        def set_to_available_state
+          broker.update_state(ServiceBrokerStateEnum::AVAILABLE)
         end
 
         def synchronizing_state
@@ -82,24 +70,6 @@ module VCAP::CloudController
 
         def failed_state
           broker.update_state(ServiceBrokerStateEnum::SYNCHRONIZATION_FAILED)
-        end
-
-        def available_state
-          broker.update_state(ServiceBrokerStateEnum::AVAILABLE)
-        end
-
-        def fail_with_invalid_catalog(errors)
-          full_message = formatter.format(errors)
-          raise CloudController::Errors::ApiError.new_from_details('ServiceBrokerCatalogInvalid', full_message)
-        end
-
-        def fail_with_incompatible_catalog(errors)
-          full_message = formatter.format(errors)
-          raise CloudController::Errors::ApiError.new_from_details('ServiceBrokerCatalogIncompatible', full_message)
-        end
-
-        def collect_warnings
-          (service_manager.warnings + client_manager.warnings).map { |w| { detail: w } }
         end
       end
     end
