+++ app/jobs/v3/services/update_broker_job.rb
@@ -43,29 +43,19 @@ module VCAP::CloudController
           @update_request = ServiceBrokerUpdateRequest.find(guid: update_request_guid)
           @broker = ServiceBroker.find(id: @update_request.service_broker_id)
           @previous_broker_state = previous_broker_state
-          @formatter = VCAP::Services::ServiceBrokers::ValidationErrorsFormatter.new
-          @service_event_repository = VCAP::CloudController::Repositories::ServiceEventRepository::WithBrokerActor.new
-          @client_manager = VCAP::Services::SSO::DashboardClientManager.new(broker, service_event_repository)
-          @service_manager = VCAP::Services::ServiceBrokers::ServiceManager.new(service_event_repository)
+          @catalog_updater = VCAP::CloudController::V3::ServiceBrokerCatalogUpdater.new(@broker)
         end
 
         def perform
           ServiceBroker.db.transaction do
             broker.update(update_params)
 
-            catalog = VCAP::Services::ServiceBrokers::V2::Catalog.new(broker, broker_client.catalog)
-
-            raise fail_with_invalid_catalog(catalog.validation_errors) unless catalog.valid?
-            raise fail_with_incompatible_catalog(catalog.incompatibility_errors) unless catalog.compatible?
-            unless client_manager.synchronize_clients_with_catalog(catalog)
-              raise fail_with_invalid_catalog(client_manager.errors)
-            end
-
-            service_manager.sync_services_and_plans(catalog)
+            @warnings = @catalog_updater.refresh
           end
 
           set_to_available_state
-          collect_warnings
+
+          @warnings
         rescue => e
           reset_to_previous_state
 
@@ -80,6 +70,10 @@ module VCAP::CloudController
 
         private
 
+        def set_to_available_state
+          broker.update_state(ServiceBrokerStateEnum::AVAILABLE)
+        end
+
         def update_params
           params = {}
           params[:name] = update_request.name unless update_request.name.nil?
@@ -92,13 +86,7 @@ module VCAP::CloudController
           params
         end
 
-        attr_reader :broker_guid, :broker,
-          :formatter, :client_manager, :service_event_repository,
-          :service_manager, :warnings, :update_request, :previous_broker_state
-
-        def broker_client
-          @broker_client ||= VCAP::Services::ServiceClientProvider.provide(broker: broker)
-        end
+        attr_reader :broker, :update_request, :previous_broker_state
 
         def reset_to_previous_state
           if previous_broker_state.nil?
@@ -106,24 +94,6 @@ module VCAP::CloudController
           else
             broker.update_state(previous_broker_state)
           end
-        end
-
-        def set_to_available_state
-          broker.update_state(ServiceBrokerStateEnum::AVAILABLE)
-        end
-
-        def fail_with_invalid_catalog(errors)
-          full_message = formatter.format(errors)
-          raise CloudController::Errors::ApiError.new_from_details('ServiceBrokerCatalogInvalid', full_message)
-        end
-
-        def fail_with_incompatible_catalog(errors)
-          full_message = formatter.format(errors)
-          raise CloudController::Errors::ApiError.new_from_details('ServiceBrokerCatalogIncompatible', full_message)
-        end
-
-        def collect_warnings
-          (service_manager.warnings + client_manager.warnings).map { |w| { detail: w } }
         end
       end
     end
